@@ -1,14 +1,14 @@
 // Angular
+import '@angular/animations';
 import '@angular/platform-browser';
 import '@angular/platform-browser-dynamic';
 import '@angular/core';
 import '@angular/common';
-import '@angular/http';
+import '@angular/common/http';
 import '@angular/router';
 import '@angular/forms';
 
 // RxJS
-// import 'rxjs';
-import {Observable} from 'rxjs/Rx';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/catch';
+import {Observable} from 'rxjs/internal/Observable';
+import {throwError} from 'rxjs/internal/observable/throwError';
+import {catchError, map, share} from 'rxjs/operators';
